version_1_wesbite_builder_simple/pyproject.toml
[project]
name = "version-1-wesbite-builder-simple"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.2.1",
]
